@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/Cambricon/cambricon-k8s-device-plugin/device-plugin/pkg/cntopo (interfaces: Cntopo)
+//
+// Generated by this command:
+//
+//	mockgen -package mock -destination pkg/cntopo/mock/cntopo.go -mock_names=Cntopo=Cntopo github.com/Cambricon/cambricon-k8s-device-plugin/device-plugin/pkg/cntopo Cntopo
+//
 
 // Package mock is a generated GoMock package.
 package mock
@@ -8,13 +13,14 @@ import (
 	reflect "reflect"
 
 	cntopo "github.com/Cambricon/cambricon-k8s-device-plugin/device-plugin/pkg/cntopo"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // Cntopo is a mock of Cntopo interface.
 type Cntopo struct {
 	ctrl     *gomock.Controller
 	recorder *CntopoMockRecorder
+	isgomock struct{}
 }
 
 // CntopoMockRecorder is the mock recorder for Cntopo.
@@ -35,16 +41,16 @@ func (m *Cntopo) EXPECT() *CntopoMockRecorder {
 }
 
 // GetRings mocks base method.
-func (m *Cntopo) GetRings(arg0 []uint, arg1 int) ([]cntopo.Ring, error) {
+func (m *Cntopo) GetRings(available []uint, size int) ([]cntopo.Ring, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetRings", arg0, arg1)
+	ret := m.ctrl.Call(m, "GetRings", available, size)
 	ret0, _ := ret[0].([]cntopo.Ring)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // GetRings indicates an expected call of GetRings.
-func (mr *CntopoMockRecorder) GetRings(arg0, arg1 interface{}) *gomock.Call {
+func (mr *CntopoMockRecorder) GetRings(available, size any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetRings", reflect.TypeOf((*Cntopo)(nil).GetRings), arg0, arg1)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetRings", reflect.TypeOf((*Cntopo)(nil).GetRings), available, size)
 }
